requirements.txt
Django==2.2.6
psycopg2
numpy
openpyxl
#GDAL
#uwsgi
#django-chart-tools
django-ckeditor
pillow
#mysqlclient
redis==2.10.6
rq
# cx_Oracle
python-dateutil
pyyaml
#ua-parser
#user-agents
#django-user-agents
